backend/requirements.txt
# Django core
Django==4.2.7
djangorestframework==3.14.0

django-filter==23.3
django-cors-headers==4.3.0
django-environ==0.11.2
dj-database-url==1.2.0
# Database
psycopg2-binary==2.9.7

# Authentication
djangorestframework-simplejwt==5.3.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# API tools
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0

# WSGI server
gunicorn==21.2.0

# Development tools
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
black==23.10.1
flake8==6.1.0
